added a change to allow multiple days in the one project
