+++ input/day1
